Fixing FastCells
